BLAN-973 - general fixes for profile page - refactoring
