+++ src/i18n/ua/errors.js
@@ -28,4 +28,5 @@ export default {
   'duration-10min-3hours': 'Допустима тривалість події від 10 хв до 3 год',
   'time-more-than-one-hour': 'Час початку має бути принаймні на одну годину пізніше поточного часу',
   'duration-must-be-round': 'Тривалість події має бути круглим числом, перевірте правильність введених даних',
+  'check-data': 'Перевірте правильність введених даних',
 }
